Fixed compilation warning: on 64bit Windows, long int is smaller than (void *)

+++ tests/test_ptr_sort2.c
@@ -24,7 +24,7 @@ static inline long int random(void) {
  * Create a random pointer
  */
 static void *random_pointer(void) {
-  return (void *) random();
+  return (void *) ((size_t) random());
 }
